app/ML/requirements.txt
chromadb==0.6.3
fastapi==0.115.12
langchain==0.3.24
langchain_chroma==0.2.3
langchain_core==0.3.55
langchain_huggingface==0.1.2
langchain_openai==0.3.14
pydantic==2.11.3
python-dotenv==1.1.0
uvicorn==0.34.2
shortuuid==1.0.13
llama-parse
PyPDF2
